Put all MetaData classes into a new module and refactored the code that needed them.
